# Associating threadpool queues with CPU cores.