@VSadov 2018-06-10
@dotnet-bot test Windows_NT x86 Checked Innerloop Build and Test please