@VSadov 2018-06-10
@dotnet-bot test this please