@VSadov 2018-06-14
@dotnet-bot test this please