@VSadov 2018-06-19
@stephentoub - nothing yet. I had an idea and tried to do some experiments.  

The idea is that a que-per-core approach (vs. current que-per-thread) could result in fewer cases of cross-core task stealing. Fewer queues would mean that we would not steal as much.
And when steals do happen, we could try to steal from the "near" cores first, if possible. That may, in theory, improve overall use of core/node/socket caches.
 
However the change would have to permit occasionally concurrent pushes/pops of the tasks and it is a bit delicate to do that both correctly and efficiently.