@VSadov 2018-11-25
@dotnet-bot test this please