@VSadov 2018-11-26
@dotnet-bot test this please