@VSadov 2018-11-29
@dotnet-bot test this please