@VSadov 2018-11-29
@dotnet-bot test Windows_NT x64 Release CoreFX Tests please