@VSadov 2018-11-30
@dotnet-bot test this please