@VSadov 2018-12-01
@dotnet-bot test Windows_NT x64 Checked CoreFX Tests please
@dotnet-bot test Windows_NT x64 Release CoreFX Tests please