@VSadov 2018-12-05
@dotnet-bot test Windows_NT x64 Release CoreFX Tests please