@VSadov 2018-12-05
On a small machine we are a bit slower. 
Especially when one worker blasts ultra-shallow/ultra-short tasks.  Need to look into that.

== On  4-core HT (8 logical) Kaby Lake 

New TP
```
Operations per second on 8 Cores
                                                                             Parallelism
                                  Serial          2x         16x         64x        512x
QUWI Queue Local (TP)            7.778 M    15.436 M    33.564 M    32.893 M    33.912 M
- Depth    2                    12.254 M    24.817 M    33.034 M    30.051 M    30.876 M
- Depth   16                    40.470 M    42.724 M    42.146 M    42.006 M    43.161 M
- Depth   64                    46.602 M    45.945 M    48.790 M    48.399 M    47.148 M
- Depth  512                    50.164 M    49.738 M    49.425 M    50.462 M    50.786 M
```

Old TP
```
Operations per second on 8 Cores
                                                                             Parallelism
                                  Serial          2x         16x         64x        512x
QUWI Queue Local (TP)            3.332 M    10.595 M    46.463 M    50.633 M    52.375 M
- Depth    2                     9.217 M    18.370 M    49.936 M    50.135 M    50.101 M
- Depth   16                    37.760 M    42.663 M    53.773 M    54.490 M    52.386 M
- Depth   64                    48.870 M    52.080 M    54.238 M    53.214 M    51.677 M
- Depth  512                    51.214 M    52.307 M    52.046 M    52.693 M    50.440 M
```