@VSadov 2018-12-05
A better story on a big(er) machine -

== On  24-core HT (48 logical)  Ivy Bridge 

New TP
```
Operations per second on 48 Cores
                                                                             Parallelism
QUWI Queue Local (TP)            2.246 M     4.300 M    25.888 M    30.773 M    33.645 M
- Depth    2                     3.979 M     8.543 M    28.738 M    31.666 M    34.559 M
- Depth   16                    24.031 M    27.997 M    36.125 M    37.067 M    36.099 M
- Depth   64                    31.527 M    33.211 M    35.756 M    36.933 M    35.569 M
- Depth  512                    34.840 M    33.912 M    36.532 M    35.159 M    34.829 M
```

Old TP
```
Operations per second on 48 Cores
                                                                             Parallelism
QUWI Queue Local (TP)          456.541 k   474.346 k    13.445 M    25.546 M    30.912 M
- Depth    2                   467.612 k   525.907 k    18.574 M    30.994 M    31.191 M
- Depth   16                    11.620 M    19.718 M    31.577 M    33.412 M    32.584 M
- Depth   64                    25.913 M    28.845 M    32.754 M    33.509 M    32.003 M
- Depth  512                    28.096 M    29.237 M    29.556 M    29.517 M    27.842 M
```