@VSadov 2018-12-05
There is also an issue with `SslStream_SameCertUsedForClientAndServer_Ok`

It looks like the test has stravation/deadlock issues. I have seen it failing on old TP as well. 
I wonder why it is more likely to see the failure with the new TP.