@VSadov 2018-12-05
These are all ideal cases though - tasks do not wait for anything, or do much work at all. Even GC is suspended while we run.

In reality threads do more work than that and often perform some measure of waiting. TP compensates by adding a few more workers.

The idea is to improve that scenario, but we do not want to regress the ideal case too much either.