@VSadov 2018-12-05
There is also an issue with `SslStream_SameCertUsedForClientAndServer_Ok`

It looks like the test has stravation/deadlock issues due to semaphore waiting in tasks. I have seen the test failing on old TP as well. 
I wonder though why it is more likely to see the failures with the new TP.