@VSadov 2018-12-05
There are lots of tests in `SslStream` area that look like:
```c#
                var tasks = new Task[2];

                tasks[0] = server.AuthenticateAsServerAsync(certificate, true, false);
                tasks[1] = client.AuthenticateAsClientAsync(
                                            certificate.GetNameInfo(X509NameType.SimpleName, false),
                                            clientCertificateCollection, false);


                await Task.WhenAll(tasks).TimeoutAfter(15 * 1000);
```

The server/client tasks communicate with each, often multiple times, other a virtual network that is essentially 2 packet queues + 2 semaphores with timeouts. 

When a bunch of tests gets scheduled there is a lot of waiting tasks. I assume the expectation is that there is no circular wait, so eventually all the tasks get to run, but depending on how they get scheduled the e2e time for each test may vary.

This particular test has a very short timeout - 15sec, most other tests have longer. I guess that is why this one tend to fail more often. 
Maybe it is just a matter of increasing the timeout. - say, if the worst case requires 30 TP threads, 15sec may not be enough to reliably pass.

Other possibility is that the test indeed has a deadlock that is triggered by a particular scheduling order, but that seems less likely. 

I have never seen this timeout to repro locally. 
There must be something with the test machines as well. Are they smaller machines - like 1/2 core VMs or something?