@VSadov 2018-12-06
@dotnet-bot test this please