@VSadov 2018-12-23
@dotnet-bot test this please