@VSadov 2018-12-24
@dotnet-bot test this please