@VSadov 2018-12-25
@dotnet-bot test this please