@VSadov 2018-12-27
@dotnet-bot test this please