@VSadov 2018-12-29
@dotnet-bot test this please