@VSadov 2018-12-30
@dotnet-bot test this please