@VSadov 2018-12-30
@dotnet-bot test OSX10.12 x64 Checked Innerloop Build and Test please