@VSadov 2019-08-15
New throughput/overhead results:

====== Master (global and local queues):
```
Testing 2,621,440 calls, with GCs after 262,144 calls.
Operations per second on 12 Cores
                                                                             Parallelism
                                  Serial          2x         16x         64x        512x
QUWI No Queues (TP)              1.631 M     5.910 M     9.111 M     9.101 M     9.321 M
- Depth    2                     5.252 M     8.463 M     8.394 M     8.490 M     8.420 M
- Depth   16                     8.409 M     8.363 M     8.373 M     8.402 M     8.417 M
- Depth   64                     8.538 M     8.519 M     8.502 M     8.500 M     8.513 M
- Depth  512                     8.579 M     8.584 M     8.575 M     8.583 M     8.920 M


QUWI Queue Local (TP)            1.521 M     5.860 M    35.595 M    48.545 M    52.873 M
- Depth    2                     4.191 M    14.324 M    44.401 M    53.220 M    53.658 M
- Depth   16                    35.904 M    44.444 M    55.769 M    56.682 M    54.123 M
- Depth   64                    48.005 M    50.327 M    56.267 M    55.770 M    53.705 M
- Depth  512                    51.953 M    54.045 M    56.088 M    55.433 M    54.566 M
```

Complete results for async/yield chains, returns and fanouts:

====== New Queue (global and local queues):
```
Testing 2,621,440 calls, with GCs after 262,144 calls.
Operations per second on 12 Cores
                                                                             Parallelism
                                  Serial          2x         16x         64x        512x
QUWI No Queues (TP)              1.903 M     5.538 M     9.707 M    10.919 M    10.845 M
- Depth    2                     9.886 M     9.833 M     9.734 M     9.882 M     9.797 M
- Depth   16                     9.842 M    10.195 M    10.235 M    10.239 M    10.259 M
- Depth   64                    10.445 M    10.460 M    10.403 M    10.462 M    10.402 M
- Depth  512                    10.501 M    10.535 M    10.526 M     9.450 M     9.695 M


QUWI Queue Local (TP)            9.710 M    21.610 M    53.522 M    59.379 M    69.616 M
- Depth    2                    17.367 M    43.470 M    72.397 M    71.998 M    69.910 M
- Depth   16                    82.986 M    84.046 M    86.097 M    84.132 M    84.295 M
- Depth   64                    87.661 M    88.506 M    86.531 M    84.830 M    85.900 M
- Depth  512                    83.351 M    83.392 M    82.440 M    83.876 M    83.804 M
```

Complete results for async/yield chains, returns and fanouts: