@kouvel 2019-09-25
`Parallel.For` from a thread pool thread leads to thread starvation
- When there is only one item in a local queue segment, `LocalQueueSegment.TryRemove()` fails to remove the item because `position == deqPosition` and the second compare-exchange always fails
- After fixing the above, there is still a new issue due to way `Parallel.For` works and due to local queue changes due to processor ID changes
  - Each task queued by `Parallel.For` queues another task and processes its batch of work, and the initial task waits for queued tasks
  - Before this change, `Task.Wait()` would be able to remove the tasks that were queued by the same thread (unless it was stolen), and this way if the queued task is not stolen, one thread is still able to execute remaining batches inline and complete the `Parallel.For`
  - After this change, since a thread may change queues due to processor ID change, it is often not able to remove a task that the same thread queued and instead blocks the thread. Eventually all threads block and leads to starvation.
- Prototyped a fix to check all items in a segment, check all segments, and to check all local queues. It's a bit unfortunate to have to check all locally queued items but it would be better than to block.
  - https://github.com/kouvel/coreclr/compare/9051819c3289eda44ba9b6cc70ecebec1156d240...0d44740821857a22a927ad846ec1ddad2b387a44

Repro:

```c#
using System;
using System.Diagnostics;
using System.Linq;
using System.Threading;
using System.Threading.Tasks;

static class ParallelForTest
{
    private static void Main()
    {
        int parallelForCount = Environment.ProcessorCount * 2;

        var threadOperationCounts = new int[(parallelForCount + 1) * 16];
        Action<int> runLoopIteration = i => { };
        WaitCallback workStart = null;
        workStart = data =>
        {
            int threadArrayIndex = ((int)data + 1) * 16;
            Parallel.For(0, 100, runLoopIteration);
            ++threadOperationCounts[threadArrayIndex];
            ThreadPool.QueueUserWorkItem(workStart, data);
        };

        for (int i = 0; i < parallelForCount; ++i)
            ThreadPool.QueueUserWorkItem(workStart, i);

        var sw = new Stopwatch();
        var operationCountDiffs = new int[parallelForCount];
        while (true)
        {
            for (int i = 0; i < parallelForCount; ++i)
                operationCountDiffs[i] = -threadOperationCounts[(i + 1) * 16];

            sw.Restart();
            Thread.Sleep(500);
            sw.Stop();

            for (int i = 0; i < parallelForCount; ++i)
                operationCountDiffs[i] += threadOperationCounts[(i + 1) * 16];

            var stalled = "";
            for (int i = 0; i < parallelForCount; ++i)
            {
                if (operationCountDiffs[i] > 0)
                    continue;

                if (stalled.Length == 0)
                    stalled += "Stalled: ";
                else
                    stalled += ' ';
                stalled += i;
            }
            if (stalled.Length > 0)
                stalled += Environment.NewLine;
            Console.WriteLine($"{(stalled.Length > 0 ? stalled : "")}Operations/ms: {operationCountDiffs.Sum() / sw.Elapsed.TotalMilliseconds,15:0.000}");
        }
    }
}
```

Before:

```
Operations/ms:        1907.421
Operations/ms:        2039.355
Operations/ms:        2060.849
Operations/ms:        2050.903
Operations/ms:        2054.450
Operations/ms:        2058.023
Operations/ms:        2056.595
Operations/ms:        2060.018
```

After:

```
Stalled: 0 2 8 9 10 11 12 13 14 15
Operations/ms:           0.012
Stalled: 1 6 9 10 11 15
Operations/ms:           8.105
Stalled: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15
Operations/ms:           0.000
Stalled: 11 15
Operations/ms:          14.043
Stalled: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15
Operations/ms:           0.000
Stalled: 0 1 2 3 5 7 8
Operations/ms:           5.117
Stalled: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15
Operations/ms:           0.000
Stalled: 0 2 3 4 5 7 8 11 14 15
Operations/ms:           0.261
Operations/ms:          12.211
Stalled: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15
Operations/ms:           0.000
Operations/ms:         129.896
Stalled: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15
Operations/ms:           0.000
Operations/ms:          46.755
Stalled: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15
Operations/ms:           0.000
Operations/ms:          25.768
Stalled: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15
Operations/ms:           0.000
Operations/ms:         450.995
Operations/ms:        1043.407
Operations/ms:        1008.231
Operations/ms:        1049.834
Operations/ms:        1186.806
Operations/ms:        1097.863
Operations/ms:        1002.073
Operations/ms:        1058.687
Operations/ms:        1123.870
```

All threads are stalled for many seconds due to starvation, even after starvation is resolved, throughput is half of baseline due to not being able to remove and process tasks inline.

With prototype fix:

```
Operations/ms:        1875.792
Operations/ms:        1985.873
Operations/ms:        1977.451
Operations/ms:        1966.705
Operations/ms:        2032.597
Operations/ms:        1974.248
Operations/ms:        2019.481
Operations/ms:        1970.111
```